refactor: split google stream adapter
